requirements.txt
Flask==2.3.3
Werkzeug==2.3.8
Pillow==10.1.0
